change AboutPageHero and mainPageHero
Я додав ваші зауваження, і також додав форму у mainPageHero. Зробив гумову верстку, та переніс весь контент у файл json
