+++ src/modules/aboutPageHero/components/AboutPageHero/aboutPageHero.module.scss
