+++ src/modules/aboutPageHero/components/infoCardHelp/infoCardHelp.jsx
@@ -1,14 +1,14 @@
-import s from './infoCardHelp.module.scss';
-import helpData from '../data/helpData.json';
 import Icon from '@/shared/components/Icon/Icon.jsx';
+import helpData from '../../data/sectionContent.json';
+import s from './infoCardHelp.module.scss';
 
 const InfoCardHelp = () => {
   return (
     <div className={s.containerCards}>
-      <h2 className={s.title}>Те, що ми вже зробили разом</h2>
+      <h2 className={s.title}>{helpData.statsTitle}</h2>
       <ul className={s.infoCards}>
-        {helpData.stats.map(({ amount, description, iconName }, index) => (
-          <li key={index} className={s.card}>
+        {helpData.stats.map(({ id, amount, description, iconName }) => (
+          <li key={id} className={s.card}>
             <Icon
               iconName={iconName}
               width="40"
