+++ src/modules/aboutPageHero/components/infoCardHelp/infoCardHelp.module.scss
@@ -39,13 +39,14 @@
   align-items: center;
   border-radius: 32px;
   padding: 16px 32px;
-  width: 370px;
+  max-width: 370px;
   background: $img-backdrop-color;
   margin-bottom: 40px;
 
   @include media(tablet) {
     padding: 40px 0;
     width: 688px;
+    max-width: none;
   }
 
   @include media(desktop) {
