--- src/modules/aboutPageHero/data/helpData.json
@@ -1,24 +1,0 @@
-{
-  "stats": [
-    {
-      "amount": "12 000+",
-      "description": "нагодованих людей",
-      "iconName": "icon-apple"
-    },
-    {
-      "amount": "4 000+",
-      "description": "забезпечені одягом",
-      "iconName": "icon-t-shirt"
-    },
-    {
-      "amount": "8 500+",
-      "description": "забезпеченно водою",
-      "iconName": "icon-water-min"
-    },
-    {
-      "amount": "3 000+",
-      "description": "отримали ліки",
-      "iconName": "icon-first-aid-kit"
-    }
-  ]
-}
