+++ src/modules/aboutPageHero/data/sectionContent.json
@@ -1,0 +1,36 @@
+{
+  "heroTitle": "Ми тут, щоб допомагати",
+  "introText": {
+    "sectionTitle": "Заради кого ми працюємо",
+    "description1": "З перших днів війни ми об’єднали наші зусилля, щоб підтримати тих, хто опинився у важких умовах.",
+    "description2": "Наша мета — допомагати кожному, хто опинився в складних життєвих умовах, аби вони знали, що про них пам’ятають і готові підтримати.",
+    "description3": "Подарувати відчуття дому, турботи та впевненості в завтрашньому дні всім, хто цього потребує — цивільним, військовим, громадам та покинутим тваринам."
+  },
+  "statsTitle": "Те, що ми вже зробили разом",
+  "stats": [
+    {
+      "id": "1",
+      "amount": "12 000+",
+      "description": "нагодованих людей",
+      "iconName": "icon-apple"
+    },
+    {
+      "id": "2",
+      "amount": "4 000+",
+      "description": "забезпечені одягом",
+      "iconName": "icon-t-shirt"
+    },
+    {
+      "id": "3",
+      "amount": "8 500+",
+      "description": "забезпеченно водою",
+      "iconName": "icon-water-min"
+    },
+    {
+      "id": "4",
+      "amount": "3 000+",
+      "description": "отримали ліки",
+      "iconName": "icon-first-aid-kit"
+    }
+  ]
+}
