+++ src/modules/mainPageHero/components/InfoCardsList/InfoCardsList.jsx
@@ -1,12 +1,12 @@
+import sectionContent from '../../data/sectionContent.json';
 import s from './InfoCardsList.module.scss';
-import infoData from '../../data/infoData.json';
 
 const InfoCardsList = () => {
   return (
     <div className={s.containerCards}>
       <ul className={s.infoCards}>
-        {infoData.map(({ amount, label }, index) => (
-          <li key={index} className={s.card}>
+        {sectionContent.infoCards.map(({ id, amount, label }) => (
+          <li key={id} className={s.card}>
             <h2 className={s.amount}>{amount}</h2>
             <p className={s.label}>{label}</p>
           </li>
