+++ src/modules/mainPageHero/components/InfoCardsList/InfoCardsList.module.scss
@@ -1,8 +1,9 @@
 @import '@/shared/styles/mixins';
 
 .containerCards {
-  width: 370px;
-  height: 344px;
+  width: 100%;
+  max-width: 370px;
+  min-height: 344px;
   padding: 32px 60px;
   border-radius: 32px;
   background-color: #dddddd;
@@ -12,11 +13,13 @@
   @include media(tablet) {
     width: 688px;
     height: 360px;
+    max-width: none;
+    min-height: auto;
     padding: 40px 40px;
     margin-top: 320px;
     display: flex;
     align-items: center;
-    top: 90%;
+    top: 105%;
   }
 
   @include media(desktop) {
