+++ src/modules/mainPageHero/components/MainPageHero/MainPageHero.module.scss
@@ -1,7 +1,8 @@
 @import '@/shared/styles/mixins';
 
 .imagePlaceholder {
-  width: 402px;
+  max-width: 402px;
+  width: 100%;
   height: 540px;
   background: #666666;
   position: relative;
@@ -12,8 +13,14 @@
   padding: 16px;
   box-sizing: border-box;
 
+  @include media(mobile) {
+    width: 100%;
+    max-width: none;
+    width: 402px;
+  }
   @include media(tablet) {
     width: 768px;
+    max-width: none;
     height: 540px;
     padding: 0 66px;
   }
@@ -27,14 +34,6 @@
   }
 }
 
-.containerTitleForm {
-  @include media(desktop) {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-}
-
 .contentContainer {
   display: flex;
   flex-direction: column;
@@ -42,7 +41,7 @@
   min-height: 1150px;
 
   @include media(tablet) {
-    min-height: 1200px;
+    min-height: 1300px;
   }
 
   @include media(desktop) {
@@ -64,8 +63,8 @@
 
   @include media(tablet) {
     @include font(36, 54);
-    width: 700px;
-    top: 30%;
+    width: 500px;
+    top: 20%;
   }
 
   @include media(desktop) {
@@ -77,29 +76,29 @@
   }
 }
 
-.formPlaceholder {
-  width: 374px;
-  height: 398px;
-  background: #f5f5f5;
-  border-radius: 32px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.form {
+  max-width: 374px;
   position: absolute;
   top: 30%;
   left: 50%;
   transform: translate(-50%, 40%);
-
+  width: calc(100% - 32px);
+  @include media(mobile) {
+    width: 100%;
+    top: 30%;
+    left: 50%;
+    transform: translate(-50%, 40%);
+  }
   @include media(tablet) {
-    width: 450px;
-    height: 406px;
-    transform: translate(-80%, 40%);
-    top: 40%;
+    max-width: none;
+    width: auto;
+    transform: translate(-76%, 40%);
+    top: 50%;
   }
 
   @include media(desktop) {
-    width: 416px;
-    height: 406px;
+    max-width: none;
+    width: auto;
     top: 10%;
     left: 83%;
     transform: translate(-50%, 40%);
