--- src/modules/mainPageHero/data/infoData.json
@@ -1,14 +1,0 @@
-[
-  {
-    "amount": "1 млн грн",
-    "label": "Загальна сума зборів"
-  },
-  {
-    "amount": "8 000+",
-    "label": "Кількість реалізованих проєктів"
-  },
-  {
-    "amount": "35 000+",
-    "label": "Кількість врятованих життів"
-  }
-]
