+++ src/modules/mainPageHero/data/sectionContent.json
@@ -1,0 +1,20 @@
+{
+  "heroTitle": "Допомагаємо тим, хто опинився в важких життєвих умовах",
+  "infoCards": [
+    {
+      "id": 1,
+      "amount": "1 млн грн",
+      "label": "Загальна сума зборів"
+    },
+    {
+      "id": 2,
+      "amount": "8 000+",
+      "label": "Кількість реалізованих проєктів"
+    },
+    {
+      "id": 3,
+      "amount": "35 000+",
+      "label": "Кількість врятованих життів"
+    }
+  ]
+}
